+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ## Bug Fixes
 
 * Fix Toodledo start & due date timezone synchronization
+* Fix create backup folder on Windows
+* Fix skip restore backup settings
 
 # 0.8.1